Move Pururin from `english` to `all`
