+++ app/src/main/java/eu/kanade/tachiyomi/source/online/all/NHentai.kt
@@ -41,7 +41,6 @@ class NHentai(delegate: HttpSource, val context: Context) :
     PagePreviewSource {
     override val metaClass = NHentaiSearchMetadata::class
     override fun newMetaInstance() = NHentaiSearchMetadata()
-    override val lang = delegate.lang
 
     private val sourcePreferences: SharedPreferences by lazy {
         context.getSharedPreferences("source_$id", 0x0000)
